frontend: implement `del` -- it was lowered as `pass`, silently
python_ifa_build_if1.cc had

    case PY_pass_stmt:
    case PY_del_stmt:
      return 0;

so EVERY `del` in every program was discarded with no diagnostic.

Found by hand-testing the corpus subset that compiles with NO warnings
and still fails. shedskin_examples/block compiles with one line of
compiler output -- no warning, no error -- and then dies without printing
a character: SIGSEGV on the C backend, "Unhandled exception:" on LLVM.
`gdb -batch -ex run -ex bt` showed 6800 identical frames of its Huffman
`iterate`, which is

    del c[0]
    root = iterate(c)

The list never shrank, `len(c) > 1` stayed true, and the stack blew.

Lowered now the way the equivalent assignment already was:

    del o[i]    ->  o.__delitem__(i)
    del o[i:j]  ->  o[i:j] = []    (__pyc_setslice__ with an empty value,
                    which is how list.__delitem__ is itself written)

`del name` and `del o.attr` stay no-ops -- neither has an effect a
compiled program can observe without a runtime binding model, and
CPython's own observable effect (a later NameError) is not modelled
either. build_syms marks only the PY_power targets PY_STORE, not
mark_store(), which would BIND a bare name instead of reading it.

The bare `del x[:]` form needed a second fix. Omitted slice bounds are
the sentinels INT_MIN/INT_MAX, and _CG_list_setslice_internal compared
them against a uint32 length -- so `l > len1` promoted INT_MIN to
2147483648, clamped the lower bound to len1, and deleted NOTHING. Made
the lengths signed; the two `if (... < 0)` branches right below exist
precisely for negative bounds, so the comparison feeding them has to be
signed too. `del x[i:j]` with explicit non-negative bounds was
unaffected, which is why only the `[:]` form exposed it.

block no longer crashes and runs to completion. It is still WRONG, for a
second unrelated reason the crash was hiding (findprobs' tuples come back
empty while len() is right) -- recorded in ifa/issues/102, not fixed
here, and not reproducible outside block's whole-program context.
linalg's `del list1[lasti:n]` also becomes real: its warnings drop 62 ->
30 and its failure moves from the BOXING refusal to a C error.

Also adds corpus_sweep.sh + sweeps/, which cache a corpus sweep per
(mode, env, working tree) so the same 15-40 minute measurement is not
repeated across sessions, and a CLAUDE.md section for it recording two
ways to get a sweep that looks real and is not (concurrent sweeps produce
spurious timeouts; `make` during a sweep makes in-flight compiles fail).

Five CI gates green; suite 302 passed / 0 failed / 14 known on both
backends.

+++ pyc_c_runtime.h
@@ -1308,7 +1308,15 @@ static inline _CG_list _CG_list_getslice_internal(_CG_list v, uint32 size, int32
 }
 
 static inline _CG_list _CG_list_setslice_internal(_CG_list l1, uint32 size, int32 l, int32 h, _CG_list l2) {
-  uint32 len1 = _CG_prim_len(0, l1), len2 = _CG_prim_len(0, l2);
+  // SIGNED. These used to be uint32, which made `l > len1` promote a
+  // negative bound to a huge unsigned value: the omitted-lower sentinel
+  // INT_MIN read as 2147483648, so `del x[:]` clamped l to len1 instead
+  // of 0 and deleted NOTHING. (`del x[i:j]` with explicit non-negative
+  // bounds was unaffected, which is why it went unnoticed -- and until
+  // `del` was lowered at all, nothing reached here.) A negative bound is
+  // the whole point of the two `if (... < 0)` branches just below, so
+  // the comparison feeding them has to be signed too.
+  int32 len1 = (int32)_CG_prim_len(0, l1), len2 = (int32)_CG_prim_len(0, l2);
   if (l > len1) l = len1;
   if (l < 0) {
     l = len1 + l;
